WIP: Non virtual AST types.
