+++ source/slang/slang-ast-type.h
@@ -13,12 +13,11 @@ class OverloadGroupType : public Type
 {
     SLANG_CLASS(OverloadGroupType)
 
-    virtual String toString() override;
-
 protected:
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
 };
 
 // The type of an initializer-list expression (before it has
@@ -27,12 +26,12 @@ class InitializerListType : public Type
 {
     SLANG_CLASS(InitializerListType)
 
-    virtual String toString() override;
-
+    
 protected:
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
 };
 
 // The type of an expression that was erroneous
@@ -40,13 +39,12 @@ class ErrorType : public Type
 {
     SLANG_CLASS(ErrorType)
 
-    virtual String toString() override;
-
 protected:
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
 // A type that takes the form of a reference to some declaration
@@ -56,9 +54,7 @@ class DeclRefType : public Type
 
     DeclRef<Decl> declRef;
 
-    virtual String toString() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
-
+    
     static RefPtr<DeclRefType> create(ASTBuilder* astBuilder, DeclRef<Decl> declRef);
 
 protected:
@@ -66,9 +62,11 @@ protected:
         : declRef(declRef)
     {}
 
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
 // Base class for types that can be used in arithmetic expressions
@@ -77,7 +75,10 @@ class ArithmeticExpressionType : public DeclRefType
     SLANG_ABSTRACT_CLASS(ArithmeticExpressionType)
 
 public:
-    virtual BasicExpressionType* GetScalarType() = 0;
+    BasicExpressionType* GetScalarType();
+
+protected:
+    BasicExpressionType* _GetScalarTypeOverride();
 };
 
 class BasicExpressionType : public ArithmeticExpressionType 
@@ -92,10 +93,11 @@ protected:
         : baseType(baseType)
     {}
 
-    virtual BasicExpressionType* GetScalarType() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-
+    
+protected:
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    BasicExpressionType* _GetScalarTypeOverride();
 };
 
 // Base type for things that are built in to the compiler,
@@ -375,13 +377,13 @@ class ArrayExpressionType : public Type
     RefPtr<Type> baseType;
     RefPtr<IntVal> arrayLength;
 
-    virtual String toString() override;
-
+    
 protected:
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
+    HashCode _getHashCodeOverride();
 };
 
 // The "type" of an expression that resolves to a type.
@@ -394,16 +396,15 @@ class TypeType : public Type
     // The type that this is the type of...
     RefPtr<Type> type;
 
-    virtual String toString() override;
-
 protected:
     TypeType(RefPtr<Type> type)
         : type(type)
     {}
 
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
 };
 
 // A vector type, e.g., `vector<T,N>`
@@ -418,10 +419,9 @@ class VectorExpressionType : public ArithmeticExpressionType
     // The number of elements
     RefPtr<IntVal> elementCount;
 
-    virtual String toString() override;
-
 protected:
-    virtual BasicExpressionType* GetScalarType() override;
+    String _toStringOverride();
+    BasicExpressionType* _GetScalarTypeOverride();
 };
 
 // A matrix type, e.g., `matrix<T,R,C>`
@@ -435,10 +435,9 @@ class MatrixExpressionType : public ArithmeticExpressionType
 
     RefPtr<Type> getRowType();
 
-    virtual String toString() override;
-
 protected:
-    virtual BasicExpressionType* GetScalarType() override;
+    String _toStringOverride();
+    BasicExpressionType* _GetScalarTypeOverride();
 
 private:
     RefPtr<Type> rowType;
@@ -465,7 +464,7 @@ class PtrTypeBase : public BuiltinType
     SLANG_CLASS(PtrTypeBase)
 
     // Get the type of the pointed-to value.
-    Type*   getValueType();
+    Type* getValueType();
 };
 
 // A true (user-visible) pointer type, e.g., `T*`
@@ -508,17 +507,17 @@ class NamedExpressionType : public Type
     DeclRef<TypeDefDecl> declRef;
     RefPtr<Type> innerType;
 
-    virtual String toString() override;
-
+    
 protected:
     NamedExpressionType(
         DeclRef<TypeDefDecl> declRef)
         : declRef(declRef)
     {}
 
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
 };
 
 // A function type is defined by its parameter types
@@ -540,13 +539,12 @@ class FuncType : public Type
     Type* getParamType(UInt index) { return paramTypes[index]; }
     Type* getResultType() { return resultType; }
 
-    virtual String toString() override;
-
 protected:
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
+    String _toStringOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
 };
 
 // The "type" of an expression that names a generic declaration.
@@ -558,17 +556,17 @@ class GenericDeclRefType : public Type
 
     DeclRef<GenericDecl> const& getDeclRef() const { return declRef; }
 
-    virtual String toString() override;
-
 protected:
     GenericDeclRefType(
         DeclRef<GenericDecl> declRef)
         : declRef(declRef)
     {}
 
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
+protected:
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
 };
 
 // The "type" of a reference to a module or namespace
@@ -580,12 +578,11 @@ class NamespaceType : public Type
 
     DeclRef<NamespaceDeclBase> const& getDeclRef() const { return declRef; }
 
-    virtual String toString() override;
-
 protected:
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;    
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();    
 };
 
 // The concrete type for a value wrapped in an existential, accessible
@@ -596,11 +593,12 @@ class ExtractExistentialType : public Type
 
     DeclRef<VarDeclBase> declRef;
 
-    virtual String toString() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
+protected:
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
     /// A tagged union of zero or more other types.
@@ -615,11 +613,12 @@ class TaggedUnionType : public Type
         ///
     List<RefPtr<Type>> caseTypes;
 
-    virtual String toString() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
+protected:
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
 class ExistentialSpecializedType : public Type 
@@ -629,11 +628,12 @@ class ExistentialSpecializedType : public Type
     RefPtr<Type> baseType;
     ExpandedSpecializationArgs args;
 
-    virtual String toString() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
+protected:
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
     /// The type of `this` within a polymorphic declaration
@@ -643,11 +643,12 @@ class ThisType : public Type
 
     DeclRef<InterfaceDecl> interfaceDeclRef;
 
-    virtual String toString() override;
-    virtual bool equalsImpl(Type* type) override;
-    virtual HashCode getHashCode() override;
-    virtual RefPtr<Type> createCanonicalType() override;
-    virtual RefPtr<Val> substituteImpl(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff) override;
+protected:
+    String _toStringOverride();
+    bool _equalsImplOverride(Type* type);
+    HashCode _getHashCodeOverride();
+    RefPtr<Type> _createCanonicalTypeOverride();
+    RefPtr<Val> _substituteImplOverride(ASTBuilder* astBuilder, SubstitutionSet subst, int* ioDiff);
 };
 
 } // namespace Slang
